+++ src/features/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "tools")]
-pub mod tools;